requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
boto3==1.34.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
jinja2==3.1.2
markupsafe==2.1.3
